Refactors Tekton Dashboard reconciler to use tektoninstallerset client
