+++ pkg/reconciler/common/nometrics.go
@@ -1,12 +1,9 @@
 /*
 Copyright 2022 The Tekton Authors
-
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
-
     http://www.apache.org/licenses/LICENSE-2.0
-
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,14 +11,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package openshiftpipelinesascode
+package common
 
 import "go.uber.org/zap"
 
 type Recorder struct {
 }
 
-func NewRecorder() (*Recorder, error) {
+func NoMetrics() (*Recorder, error) {
 	return &Recorder{}, nil
 }
