+++ pkg/reconciler/kubernetes/tektonconfig/extension/dashboard_test.go
@@ -41,9 +41,9 @@ func TestEnsureTektonDashbordExists(t *testing.T) {
 	util.AssertEqual(t, err, v1alpha1.RECONCILE_AGAIN_ERR)
 
 	// during second invocation instance exists but waiting on dependencies (pipeline, triggers)
-	// hence returns DEPENDENCY_UPGRADE_PENDING_ERR
+	// hence returns RECONCILE_AGAIN_ERR
 	_, err = EnsureTektonDashboardExists(ctx, c.OperatorV1alpha1().TektonDashboards(), tConfig)
-	util.AssertEqual(t, err, v1alpha1.DEPENDENCY_UPGRADE_PENDING_ERR)
+	util.AssertEqual(t, err, v1alpha1.RECONCILE_AGAIN_ERR)
 
 	// make upgrade checks pass
 	makeUpgradeCheckPass(t, ctx, c.OperatorV1alpha1().TektonDashboards())
